backend/requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.4.4
Flask-CORS==4.0.0
Flask-SocketIO==5.3.4
python-socketio==5.9.0
python-engineio==4.7.1
psycopg2-binary==2.9.7
SQLAlchemy==2.0.21
Werkzeug==2.3.7
python-dotenv==1.0.0
stripe>=9.0.0
openai>=1.0.0
requests==2.31.0
gunicorn==21.2.0
gevent==23.9.1
gevent-websocket==0.10.1
pydantic==2.5.0
python-dateutil==2.8.2
